+++ src/application/clothingModel/Chest.java
@@ -35,7 +35,7 @@ public class Chest implements BodyPart {
     }
 
     @Override
-    public int getMeasurement(ClothingType clothingType) throws InvalidSizeException {
+    public double getMeasurement(ClothingType clothingType) throws InvalidSizeException {
         if(clothingType == ClothingType.TOP){
             return getTopSize();
         }
