+++ src/application/clothingModel/Waist.java
@@ -57,7 +57,7 @@ public class Waist implements BodyPart {
     }
 
     @Override
-    public int getMeasurement(ClothingType clothingType) throws InvalidSizeException {
+    public double getMeasurement(ClothingType clothingType) throws InvalidSizeException {
         if(clothingType == ClothingType.TOP && !this.male){
             return getFemaleTopSize();
         }else if(clothingType == ClothingType.BOTTOM){
